requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain
langchain_community
openai-whisper
ffmpeg-python